Moved user agent generator to config from utils
