+++ src/facebook/config.py
@@ -1,5 +1,7 @@
-from dotenv import load_dotenv
 import os
+
+from fake_useragent import UserAgent
+from dotenv import load_dotenv
 
 load_dotenv()
 
@@ -10,3 +12,10 @@ def get_proxy_link() -> str:
         return proxy_ulr
     else:
         raise ValueError("The proxy value can't be empty!")
+
+
+class FakeUserAgent:
+    @staticmethod
+    def get_fake_ua() -> str:
+        ua = UserAgent(browsers="Firefox", os="Windows")
+        return ua.random
